+++ spec/integration/resource/api500/synergy/sas_logical_interconnect_group/delete_spec.rb
@@ -11,9 +11,8 @@
 
 require 'spec_helper'
 
-klass = OneviewSDK::API300::Synergy::SASLogicalInterconnectGroup
-RSpec.describe klass, integration: true, type: CREATE, sequence: seq(klass) do
-  let(:current_client) { $client_300_synergy }
-
-  include_examples 'SASLogInterGroupCreateExample', 'integration api300 context'
+klass = OneviewSDK::API500::Synergy::SASLogicalInterconnectGroup
+RSpec.describe klass, integration: true, type: DELETE, sequence: rseq(klass) do
+  let(:current_client) { $client_500_synergy }
+  include_examples 'SASLogInterGroupDeleteExample', 'integration api500 context'
 end